2DSEG_DI/2D-gaussian/requirements.txt
plyfile
tqdm
open3d
opencv-python
tensorboard
protobuf==4.25.3
torchmetrics
wis3d
trimesh
